+++ .idea/modules.xml
@@ -1,0 +1,11 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<project version="4">
+  <component name="ProjectModuleManager">
+    <modules>
+      <module fileurl="file://$PROJECT_DIR$/bv-java-grupa-2/Homework1/Homework1.iml" filepath="$PROJECT_DIR$/bv-java-grupa-2/Homework1/Homework1.iml" />
+      <module fileurl="file://$PROJECT_DIR$/bv-java-grupa-2/Homework2/Homework2.iml" filepath="$PROJECT_DIR$/bv-java-grupa-2/Homework2/Homework2.iml" />
+      <module fileurl="file://$PROJECT_DIR$/bv-java-grupa-2/Homework3/Homework3.iml" filepath="$PROJECT_DIR$/bv-java-grupa-2/Homework3/Homework3.iml" />
+      <module fileurl="file://$PROJECT_DIR$/.idea/scoala-informala-homework.iml" filepath="$PROJECT_DIR$/.idea/scoala-informala-homework.iml" />
+    </modules>
+  </component>
+</project>
